worker/requirements.txt
huggingface_hub>=0.15.1
python-dotenv>=0.21.0